pages refactor
