+++ frontend/src/pages/MealPlan/Planner.vue
@@ -87,9 +87,9 @@
 <script>
 import api from "@/api";
 import utils from "@/utils";
-import NewMeal from "../components/MealPlan/MealPlanNew";
-import EditPlan from "../components/MealPlan/MealPlanEditor";
-import ShoppingListDialog from "../components/MealPlan/ShoppingListDialog";
+import NewMeal from "@/components/MealPlan/MealPlanNew";
+import EditPlan from "@/components/MealPlan/MealPlanEditor";
+import ShoppingListDialog from "@/components/MealPlan/ShoppingListDialog";
 
 export default {
   components: {
