+++ frontend/src/pages/Recipes/AllRecipes.vue
@@ -13,8 +13,8 @@
 </template>
 
 <script>
-import CardSection from "../components/UI/CardSection";
-import CategorySidebar from "../components/UI/CategorySidebar";
+import CardSection from "@/components/UI/CardSection";
+import CategorySidebar from "@/components/UI/CategorySidebar";
 export default {
   components: {
     CardSection,
