+++ frontend/src/pages/Recipes/CategoryPage.vue
@@ -14,8 +14,8 @@
 
 <script>
 import api from "@/api";
-import CardSection from "../components/UI/CardSection";
-import CategorySidebar from "../components/UI/CategorySidebar";
+import CardSection from "@/components/UI/CardSection";
+import CategorySidebar from "@/components/UI/CategorySidebar";
 export default {
   components: {
     CardSection,
